requirements.txt
flask==3.0.2
gunicorn==21.2.0
pyautogen>=0.2.7
openai>=1.0.0
# openai==1.12.0
python-dotenv==1.0.0
uvicorn==0.27.1
hypercorn==0.15.0
requests==2.31.0 
sseclient-py==1.8.0
authlib==1.3.0
tiktoken==0.12.0
python-dateutil==2.8.2
numpy==1.26.4
pandas==2.2.1
oracledb==2.0.1
cryptography==42.0.5
sseclient-py==1.8.0
authlib==1.3.0